pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-limiter>=4.0.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
    "replit>=4.1.2",
]
